機能追加、リファクタリング 　・ElementEditorの起動にローカルファイルモードを追加（前提としてインデックス.dbと要素のスプリットxmlが作成済） 　・BDFileReader.voやutilを参照している部分をArtifactFileAccessor系に移動（ElementEditorだけ） 　・avalonEditのコード補完機能に本格対応。 　　一旦 $ と . に対応ロジックを入れたが、 $ はクラス・メソッド名の全検索時だけでいいような気がする

+++ BehaviorDevelop/ElementEditor/ElementForm.cs
@@ -14,10 +14,10 @@ using System.Collections.Generic;
 using System.Windows.Forms;
 using System.Windows.Forms.Integration;
 
-using BDFileReader.util;
-using BDFileReader.vo;
-using BDFileReader.writer;
-using BDFileReader.reader;
+using ArtifactFileAccessor.util;
+using ArtifactFileAccessor.vo;
+using ArtifactFileAccessor.writer;
+using ArtifactFileAccessor.reader;
 
 
 namespace ElementEditor
@@ -725,8 +725,12 @@ namespace ElementEditor
 
         private void ElementForm_FormClosed(object sender, FormClosedEventArgs e)
         {
-            MainForm main = (MainForm)this.Owner;
-            main.deleteOpenedElement(myElement);
+            if( this.Owner != null )
+            {
+                MainForm main = (MainForm)this.Owner;
+                main.deleteOpenedElement(myElement);
+            }
+
         }
     }
 }
